requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.6.0
pymongo==4.9.1
python-multipart==0.0.9
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
